requirements.txt
yfinance>=0.2.36
pandas==2.1.4
requests==2.31.0
python-dotenv==1.0.0
requests-cache==1.1.1
